+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.0
+
+- Support user password in the env config.
+- When `dbName` is not specified, the database will be selected by user name.
+
 ## 1.1.0
 
 - Clean command.